Handle find previously existing sourcing location by georegionid to avoid deleting a georegion previously created and validated

+++ api/src/modules/geo-coding/strategies/point-of-production.geocoding.service.ts
@@ -40,7 +40,13 @@ export class PointOfProductionGeocodingStrategy extends BaseStrategy {
           )
         ).adminRegionId;
       } catch (e) {
-        await this.geoRegionService.remove(geoRegionId as unknown as string);
+        const existingSourcingLocation: SourcingLocation | undefined =
+          await this.findExistingSourcingLocationByGeoRegionId(
+            geoRegionId as unknown as string,
+          );
+        if (!existingSourcingLocation) {
+          await this.geoRegionService.remove(geoRegionId as unknown as string);
+        }
         throw e;
       }
